refactor: encapsulate as method
